ref(dynamic-sampling): refactor ds related utils
Extracting three utility functions from dynamic-sampling/tasks/utils to dynamic-sampling/utils.
Using `has_dynamic_sampling` and `has_custom_dynamic_sampling` instead of `features.has` where appropiate.
