+++ src/sentry/dynamic_sampling/tasks/boost_low_volume_transactions.py
@@ -42,10 +42,9 @@ from sentry.dynamic_sampling.tasks.task_context import DynamicSamplingLogState, 
 from sentry.dynamic_sampling.tasks.utils import (
     dynamic_sampling_task,
     dynamic_sampling_task_with_context,
-    has_dynamic_sampling,
-    is_project_mode_sampling,
     sample_function,
 )
+from sentry.dynamic_sampling.utils import has_dynamic_sampling, is_project_mode_sampling
 from sentry.models.options.project_option import ProjectOption
 from sentry.models.organization import Organization
 from sentry.sentry_metrics import indexer
